+++ dashboard/webapp-backend/src/main/java/org/oransc/ric/portal/dashboard/exception/StatsManagerException.java
@@ -2,14 +2,14 @@
  * ========================LICENSE_START=================================
  * O-RAN-SC
  * %%
- * Copyright (C) 2019 AT&T Intellectual Property
+ * Copyright (C) 2019 - 2020 AT&T Intellectual Property
  * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,37 +17,18 @@
  * limitations under the License.
  * ========================LICENSE_END===================================
  */
- .stats__section {
-}
+package org.oransc.ric.portal.dashboard.exception;
 
-.stats-table {
-  width: 100%;
-  min-height: 100px;
-  margin-top: 10px;
-  background-color:transparent;
-}
 
-.bar-chart-card {
-    height: 100%;
-    width: 100%;
-    max-width: 400px;
-    margin-left: 1%
-}
+public class StatsManagerException extends Exception {
+	
+	private static final long serialVersionUID = -3887065423512216019L;
 
-.mat-slider {
-  width: 300px;
-}
+	public StatsManagerException() {
+		super();
+	}
 
-.mat-grid-tile {
-  background: transparent;
-}
-
-iframe {
-  width: 500px;
-  height: 420px;
-  border: 1px solid black;
-}
-
-.display-none {
-  display: none;
+	public StatsManagerException(String s) {
+		super(s);
+	}
 }